Use semantic html elements for accessibility
We received an accessibility audit for one of our projects using
react-data-table-component. Although this library uses the `role`
attribute appropriately, there's no reason for it to be using `div`
instead of `table`. This would really help screen readers know what
to do with these tables.

Does a change like this make sense?